ai_service/requirements.txt
fastapi
uvicorn[standard]
transformers
sentencepiece
accelerate
pydantic
python-dotenv
protobuf
